chore: clean up xverse api client in prep for auth [ENG-5840] (#351)
* Refactor seed vault and add key value vault

* Fix address context to use new wallet ID field

* Start migrations

* Add migration from seed vault

* Add vault specific migrations

* Fix txn context tests

* comment out seed vault tests

* Clean

* Add missing export

* Change xprivKey to seed

* Fix getWalletIds return type

* Add get wallet rootnode to seed vault

* Change wallet type to derivation type

* Change all references of seed phrase to root node and remove bip32 and bip39 libs

* Fix method call

* Ensure seed vault returns a wallet

* Add derivation params

* Change account to accountIndex

* Fix arg name

* Fix tests

* Move bip32 to peer deps

* Clean up and fix seed vault migration and types

* Rename type

* More clean up

* Clean up ts ignore

* Fix build

* Return derivation type with root node

* Ensure vault is locked if trying to unlock an unlocked vault with incorrect password

* Remove more mnemonics from memory

* rename getPaymentAccountSummaryForSeedPhrase

* Add derivation type to stx methods

* Add is initialised methods to vaults

* Add mnemonicToRootNode helper method

* Change interfaces to be consistent

* Change restoreWalletWithAccounts args type

* Clean up on migration failure

* Fix circular failure

* Don't use master vault in migrations

* Fix vault unlock and migration logic

* Make restoreVault public

* Add missing await

* Fix restore account regen logci

* Keep data encryption keys in memory and make KV vault encrypt per key

* Add wallet id to software accounts

* Keep data keys in memory

* Fix account type

* Convert key value vault to single stored item per key

* Ensure vault tries to unlock if unlocked already

* Touch up

* Remove key value vault is initialised method

* Remove unused arg

* Add soft lock and make key value vault remain unlocked per session

* Fix embellished account method

* Add encryption vault tests

* Fix typing

* More tests and fixes

* Add master vault e2e tests

* Add migration test

* Fix tests

* vault migration fixes and tests

* Make migration a bit more resilient

* Enforce key value vault allowed keys

* Clean up Xverse API deprecated methods

* Add concept of primary wallet

* Refactor deriveStxAddressChain and fix test build

* Move over to crypto.randomUUID()

* Bump github workflows to node 22

* Hide the branding property in walletId

* Add missing awaits

* Change mutex to use runExclusive

* Update api/stacksSponsor.ts


---------
