+++ api/index.ts
@@ -5,9 +5,10 @@ export { getNetworkURL, parseStxTransactionData } from './helper';
 export { default as mempoolApi } from './mempool';
 export * from './ordinals';
 export * from './ordinals/provider';
+export * from './ordinalsService';
 export * from './runes/provider';
 export * from './stacks';
+export * from './stacksApi';
+export * from './stacksSponsor';
 export * from './utxoCache';
 export * from './xverse';
-export * from './ordinalsService';
-export * from './stacksApi';
